+++ lib/homepage.dart
@@ -3,13 +3,13 @@ import 'dart:async';
 import 'package:calibre_carte/providers/color_theme_provider.dart';
 import 'package:calibre_carte/providers/update_provider.dart';
 import 'package:calibre_carte/screens/settings_screen.dart';
+import 'package:calibre_carte/widgets/books_view_widgets/books_view.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/painting.dart';
 import 'package:provider/provider.dart';
 import 'package:rxdart/rxdart.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'widgets/Books View Widgets/books_view.dart';
 
 class MyHomePage extends StatefulWidget {
   @override
