+++ lib/screens/book_details_screen.dart
@@ -14,8 +14,8 @@ import 'package:calibre_carte/models/comments.dart';
 import 'package:calibre_carte/models/data.dart';
 import 'package:calibre_carte/models/publishers.dart';
 import 'package:calibre_carte/models/ratings.dart';
-import 'package:calibre_carte/widgets/Details%20Screen%20Widgets/details_lefttile.dart';
-import 'package:calibre_carte/widgets/Details%20Screen%20Widgets/details_sidebar.dart';
+import 'package:calibre_carte/widgets/details_screen_widgets/details_lefttile.dart';
+import 'package:calibre_carte/widgets/details_screen_widgets/details_sidebar.dart';
 import 'package:flutter/material.dart';
 import '../helpers/books_provider.dart';
 import '../models/books.dart';
