+++ lib/screens/settings_screen.dart
@@ -8,10 +8,10 @@ import 'package:calibre_carte/screens/about_us.dart';
 import 'package:calibre_carte/screens/instructions_screen.dart';
 import 'package:calibre_carte/screens/license.dart';
 import 'package:calibre_carte/screens/privacy_policy.dart';
-import 'package:calibre_carte/widgets/Settings%20Screen%20Widgets/cloud_settings.dart';
-import 'package:calibre_carte/widgets/Settings%20Screen%20Widgets/dark_mode_toggle.dart';
-import 'package:calibre_carte/widgets/Settings%20Screen%20Widgets/directorychange_button.dart';
-import 'package:calibre_carte/widgets/Settings%20Screen%20Widgets/search_dropdown.dart';
+import 'package:calibre_carte/widgets/settings_screen_widgets/cloud_settings.dart';
+import 'package:calibre_carte/widgets/settings_screen_widgets/dark_mode_toggle.dart';
+import 'package:calibre_carte/widgets/settings_screen_widgets/directorychange_button.dart';
+import 'package:calibre_carte/widgets/settings_screen_widgets/search_dropdown.dart';
 import 'package:directory_picker/directory_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/painting.dart';
