+++ lib/widgets/details_screen_widgets/animated_progressbar.dart
