+++ lib/widgets/details_screen_widgets/open_format_dialog.dart
