+++ lib/widgets/details_screen_widgets/rating.dart
