+++ lib/widgets/details_screen_widgets/select_format_dialog.dart
