+++ lib/widgets/settings_screen_widgets/dropbox_dropdown.dart
@@ -9,9 +9,9 @@ import 'package:calibre_carte/helpers/metadata_cacher.dart';
 import 'package:calibre_carte/providers/color_theme_provider.dart';
 import 'package:calibre_carte/providers/update_provider.dart';
 import 'package:calibre_carte/screens/dropbox_signin_screen.dart';
-import 'package:calibre_carte/widgets/Settings%20Screen%20Widgets/connect_button.dart';
-import 'package:calibre_carte/widgets/Settings%20Screen%20Widgets/logout_button.dart';
-import 'package:calibre_carte/widgets/Settings%20Screen%20Widgets/refresh_button.dart';
+import 'package:calibre_carte/widgets/settings_screen_widgets/connect_button.dart';
+import 'package:calibre_carte/widgets/settings_screen_widgets/logout_button.dart';
+import 'package:calibre_carte/widgets/settings_screen_widgets/refresh_button.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:provider/provider.dart';
