+++ lib/widgets/settings_screen_widgets/logout_button.dart
